Use snapcraft as a snap inside snapd and not as a standalone app. According to official documentation:https://snapcraft.io/docs/build-snaps/trusty
"In the near future, snapcraft will be offered as a snap, giving you the same experience regardless of Linux distribution or release. Until then, you can build snaps using LXD or Docker."

Under Linux Mint, on the latest snapcraft standalone version, snapcraft could be used only in conjunction with LXD, and since it is under development to fully support snapcraft inside snapd, regardless of the Linux distribution, we considered to make a step ahead to adapt our documentation.

Signed-off-by: Flavian Manea <flavian.manea@rinftech.com>